# api/openapi.yaml
openapi: 3.1.0
info:
  title: Andronoma Platform API
  version: 0.1.0
  description: |
    Orchestration surface for the Andronoma campaign pipeline. All endpoints are
    versionless for now and secured via bearer tokens that map to session
    records. The service exposes helpers for discovering the pipeline ordering,
    launching runs, inspecting telemetry, and streaming structured logs.
servers:
  - url: http://localhost:8000
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /runs:
    get:
      tags: [runs]
      summary: List runs owned by the authenticated user
      responses:
        '200':
          description: Collection of pipeline runs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunListResponse'
    post:
      tags: [runs]
      summary: Create a new pipeline run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCreateRequest'
      responses:
        '201':
          description: Run record created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
  /runs/{run_id}:
    get:
      tags: [runs]
      summary: Retrieve details for a pipeline run
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The requested run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '404':
          description: Run not found or forbidden.
  /runs/{run_id}/start:
    post:
      tags: [runs]
      summary: Transition a run into the active state and enqueue workers
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Run state after attempting to start.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '400':
          description: Run is already active.
        '403':
          description: Forbidden for current user.
        '404':
          description: Run not found.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
  schemas:
...
    PipelineConfig:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Summer launch sprint
        objectives:
          type: array
          items:
            type: string
          example:
            - Validate positioning against core personas
            - Launch creative exploration sprint
        target_markets:
          type: array
          items:
            type: string
          example:
            - US
            - CA
        metadata:
          type: object
          additionalProperties: {}
    RunCreateRequest:
      type: object
      required: [config]
      properties:
        config:
          $ref: '#/components/schemas/PipelineConfig'
        budgets:
          type: object
          additionalProperties:
            type: number
            format: float
    RunListResponse:
      type: object
      required: [runs]
      properties:
        runs:
          type: array
          items:
            $ref: '#/components/schemas/RunResponse'
    RunResponse:
      type: object
      required:
        - id
        - status
        - input_payload
        - budgets
        - telemetry
        - created_at
        - updated_at
        - stages
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/RunStatus'
        input_payload:
          type: object
          additionalProperties: {}
        budgets:
          type: object
          additionalProperties:
            type: number
        telemetry:
          type: object
          additionalProperties: {}
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        stages:
          type: array
          items:
            $ref: '#/components/schemas/StageTelemetry'
    StageTelemetry:
      type: object
      required:
        - name
        - status
        - telemetry
        - notes
      properties:
        name:
          type: string
        status:
          $ref: '#/components/schemas/StageStatus'
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        telemetry:
          type: object
          additionalProperties: {}
        budget_spent:
          type: number
        notes:
          type: string
          example: Awaiting QA signal
    RunStatus:
      type: string
      enum: [pending, running, completed, failed, cancelled]
    StageStatus:
      type: string
      enum: [pending, running, completed, failed, skipped]